# docs/swagger.yaml
basePath: /api
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
...
  models.ChangeUserRoleInput:
    properties:
      role:
        example: team_leader
        type: string
    type: object
...
  models.CreateUserInput:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      created_at:
        type: string
      error:
        type: string
    type: object
...
  models.Role:
    enum:
    - super_admin
    - team_leader
    - team_member
    type: string
    x-enum-varnames:
    - SuperAdminRole
    - TeamLeaderRole
    - TeamMemberRole
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
  models.UserShortResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: RESTful API for signing and verifying files with digital signatures
  title: Tawtheeq API
  version: "1.0"
paths:
...
  /myself/name:
    put:
      consumes:
      - application/json
      description: Allows user to change their own name
      parameters:
      - description: New full name
        in: query
        name: full_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      - JWT: []
      - ApiKeyAuth: []
      summary: Change user name
      tags:
      - myself
  /myself/password:
    put:
      consumes:
      - application/json
      description: Allows user to change their own password
      parameters:
      - description: Old password
        in: query
        name: old_password
        required: true
        type: string
      - description: New password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      - JWT: []
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - myself
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserShortResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      - JWT: []
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Create User Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserShortResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Remove user
      tags:
      - users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: Allows Super Admin to change a user's role
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ChangeUserRoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Change user role
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"